add meta tag on show
