+++ app/javascript/packs/application.js
@@ -1,8 +1,10 @@
 import "bootstrap";
 import { displayRoute } from '../plugins/displaySegment';
 import { fixFooter } from '../plugins/footer';
+import { navbarShow } from '../plugins/navbarShow';
 import { navbar } from '../plugins/navbar';
 
 fixFooter();
 displayRoute();
 navbar();
+navbarShow();
