+++ app/views/itineraries/show.html.erb
@@ -1,3 +1,6 @@
+<% # Meta information for the <head> section %>
+<% title @itinerary.name %>
+
 <div id="carouselExampleControls" class="carousel slide" data-touch="false" data-ride="carousel" data-interval="false">
   <div class="carousel-inner">
     <div
